# [DRAFT] Bring back cpu capability trampoline function on the C-level for static fips build