### Issues:

CryptoAlg-1577

### Description of changes: 

The FIPS static builds fix-up `OPENSSL_ia32cap_P` by adding a trampoline label storing an offset to it. This requires addition instructions, in turn, saving cpu flag state. Otherwise, we could break soundness/correctness since the fix-up is really direct assembly code modifications without context-awareness.

We could use the build-in delocator storage objects, but they are lazily initialised, and requires branch checks for every invocation. cpu capability checks are almost always on very hot code paths. So, we want to avoid extra work.

This recovers some performance for the static fips build compared to the shared fips build.

```
# with change static
$ ./perf_aes_set_encrypt_key_static 
iterate_aes_set_encrypt_key took about 3.52033 seconds

$ LD_LIBRARY_PATH=$(pwd)/awslc-build/crypto ./perf_aes_set_encrypt_key_shared
iterate_aes_set_encrypt_key took about 1.79280 seconds

# without change
$ ./perf_aes_set_encrypt_key_static 
iterate_aes_set_encrypt_key took about 4.19714 seconds

$ LD_LIBRARY_PATH=$(pwd)/awslc-build/crypto ./perf_aes_set_encrypt_key_shared
iterate_aes_set_encrypt_key took about 1.79277 seconds
```

### Testing:
How is this change tested (unit tests, fuzz tests, etc.)? Are there any testing steps to be verified by the reviewer?

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license and 
the ISC license.