@torben-hansen 2023-03-16
> Missed some:
grep --exclude=\*.S --exclude=\*.pl -r "OPENSSL_ia32cap_P" ./crypto                                         8:35
./crypto/fipsmodule/cpucap/cpucap.c:HIDDEN uint32_t OPENSSL_ia32cap_P[4] = {0};
./crypto/fipsmodule/cpucap/internal.h:// OPENSSL_ia32cap_P contains the Intel CPUID bits when running on an x86 or
./crypto/fipsmodule/cpucap/internal.h:extern uint32_t OPENSSL_ia32cap_P[4];
./crypto/fipsmodule/cpucap/internal.h:  return OPENSSL_ia32cap_P;
./crypto/fipsmodule/cpucap/cpu_intel.c:extern uint32_t OPENSSL_ia32cap_P[4];
./crypto/fipsmodule/cpucap/cpu_intel.c:  OPENSSL_ia32cap_P[0] = edx;
./crypto/fipsmodule/cpucap/cpu_intel.c:  OPENSSL_ia32cap_P[1] = ecx;
./crypto/fipsmodule/cpucap/cpu_intel.c:  OPENSSL_ia32cap_P[2] = extended_features[0];
./crypto/fipsmodule/cpucap/cpu_intel.c:  OPENSSL_ia32cap_P[3] = extended_features[1];
./crypto/fipsmodule/cpucap/cpu_intel.c:  // The first value determines OPENSSL_ia32cap_P[0] and [1]. The second [2]
./crypto/fipsmodule/cpucap/cpu_intel.c:  handle_cpu_env(&OPENSSL_ia32cap_P[0], env1);
./crypto/fipsmodule/cpucap/cpu_intel.c:    handle_cpu_env(&OPENSSL_ia32cap_P[2], env2 + 1);

Don't replace these:
* `OPENSSL_ia32cap_get()` returns a constant pointer...
* This code is only called once, so cost is suuuuper negligible.